Readme.md
This data is a part of the Dataquest "Data Cleaning with R" project. You can find the data description and given 2 data frames - "Masterfile11_d75_final.txt and Masterfile11_gened_final" in 2 formats, excel and txt formats. We are extracting the data from these 2 files that is necessary for us and then combining these 2 data frames. 
 After that, in order to make comparisons we combine the "combined" file with the combined version of these 2 data frames as all of them have common "DBN" numbers. 
 Then we need to look at matrix correlations of "avg_sat_score" and other relevant performance measures. We could also look at the survey reponses of "parent", "student" and "teacher" variables by using "string functions" to create a new column for these responese and the relative score variables to analyze the relationship in a boxplot graph.